+++ OpenBudgeteer.Blazor/Startup.cs
@@ -4,7 +4,6 @@ using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using Blazor.FileReader;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Hosting;
@@ -15,6 +14,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using OpenBudgeteer.Core.Common;
 using OpenBudgeteer.Core.ViewModels;
+using Tewr.Blazor.FileReader;
 
 namespace OpenBudgeteer.Blazor
 {